집계탭 거래목록 테이블 클릭 시 차트 표시 오류 재수정

+++ utility/chart.py
@@ -224,7 +224,7 @@ class Chart:
                     return cgtime_
 
                 for index in df.index:
-                    cgtime = float(df['체결시간'][index] if is_tick else str(df['체결시간'][index])[:12])
+                    cgtime = int(df['체결시간'][index] if is_tick else str(df['체결시간'][index])[:12])
                     cgtime = get_cgtime(cgtime)
                     if market in (1, 3):
                         if df['주문구분'][index] == '매수':
